docs: fix duplicate declarations, clean up compiler blockers, and add master self-sufficiency plan
Deduplicates module imports and definitions across:
- src/automation/mod.rs
- src/resilience/mod.rs
- src/shell/mod.rs
- src/filesystem/mod.rs

Also adds the comprehensive SOVEREIGN_OS_OMNIPRESENT_ULTIMATE_SELF_SUFFICIENCY_PLAN.md detailing native Safe Rust replacements for over 250+ third-party tools, databases, codecs, networks, and simulators.

+++ src/security/audit.rs
@@ -65,7 +65,7 @@ impl AuditEvent for SimpleAuditEvent {
         self.id
     }
     fn event_type(&self) -> EventType {
-        unsafe { core::mem::transmute(self.event_type.load(Ordering::SeqCst)) }
+        unsafe { core::mem::transmute(self.event_type.load(Ordering::SeqCst) as u32) }
     }
     fn timestamp(&self) -> u64 {
         self.timestamp.load(Ordering::SeqCst) as u64
