+++ src/security/integrity.rs
@@ -129,7 +129,7 @@ impl SimpleFile {
     }
 
     pub fn get_status(&self) -> IntegrityStatus {
-        unsafe { core::mem::transmute(self.status.load(Ordering::SeqCst)) }
+        unsafe { core::mem::transmute(self.status.load(Ordering::SeqCst) as u32) }
     }
 
     pub fn set_status(&self, status: IntegrityStatus) {
